requirements.txt
absl-py==1.4.0
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
fastapi==0.91.0
fonttools==4.38.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
ipython==8.10.0
jedi==0.18.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.15.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.24.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.4.0
ultralytics==8.0.20
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.13.0
